// src/@types/modules/app-version/domain/models/app-version.model.d.ts
export interface AppVersionModel {
  id: string;
  platform: string;
  versionName: string;
  buildNumber: number;
  minSupportedBuild: number;
  isActive: boolean;
  storeAppId: string;
  releaseNotes: object;
  isMaintenance: boolean;
  maintenanceMessage: object;
  createdAt: Date;
  updatedAt: Date;
}